+++ src/engine/GPURenderer/r_main.cpp
@@ -2025,7 +2025,7 @@ void R_RenderPshadowMaps( const refdef_t* fd )
                 {
                     // FIXME: never actually tested this
                     mdrHeader_t* header = ( mdrHeader_t* )model->modelData;
-                    sint frameSize = ( size_t )( &( ( mdrFrame_t* )0 )->bones[ header->numBones ] );
+                    sint frameSize = ( uint64 )( &( ( mdrFrame_t* )0 )->bones[ header->numBones ] );
                     mdrFrame_t* frame = ( mdrFrame_t* )( ( uchar8* ) header + header->ofsFrames + frameSize * ent->e.frame );
                     
                     radius = frame->radius;
